building_ai_agents_using_langgraph_windows/building_adaptive_rag_agent/requirements.txt
beautifulsoup4
tiktoken
langchainhub
tavily-python
python-dotenv
pytest
langchain-tavily
faiss-cpu
ipykernel
ipython
langchain-classic
langchain-community
langchain-core
langchain-google-genai
langchain-text-splitters
numpy
pandas
matplotlib-inline
pydantic
pydantic-settings
scikit-learn
pypdf
flax